+++ source/detect_audio_language/init.d/install_whisper.sh
@@ -12,7 +12,7 @@ fi
 mpy=$(python3 -m pip list | grep moviepy)
 if [ ! "$mpy" ]; then 
     echo "**** Installing moviepy ****"
-    python3 -m pip install moviepy==2.0.0
+    python3 -m pip install moviepy==1.0.3
 else
     echo "**** moviepy already installed ****"
 fi