+++ source/detect_audio_language/changelog.md
@@ -1,4 +1,7 @@
 
+**<span style="color:#56adda">0.0.15</span>**
+- change init.d script to install correct version of moviepy
+
 **<span style="color:#56adda">0.0.14</span>**
 - revert to moviepy 1.0.3